+++ tests/testthat/test-collate.R
@@ -47,4 +47,9 @@ test_that("DESCRIPTION file is re-written only if collate changes", {
 
  })
 
-
+test_that("Collate with one file name per line (#790)", {
+  expect_equal(
+    roxygen2:::package_files('testthat/testCollateParse'),
+    c("testthat/testCollateParse/R/b.r,", "testthat/testCollateParse/R/c.r")
+  )
+})