+++ NEWS.md
@@ -1,6 +1,7 @@
 # roxygen2 6.1.0.9000
 
-* roxygen2 now specifically imports desc >= 1.2.0 (@crsh, #773, #777, #779)
+* roxygen2 now specifically imports desc >= 1.2.0 (@crsh, #773, #777, #779).
+* Correctly parse multi-line DESCRIPTION collate directives (@brodieG, #790).
 
 # roxygen2 6.1.0